Refactor Duke to Datuk
