+++ src/main/java/datuk/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package datuk;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
