+++ src/main/java/datuk/Parser.java
@@ -1,4 +1,4 @@
-package duke;
+package datuk;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
