+++ src/main/java/datuk/task/Deadline.java
@@ -1,4 +1,4 @@
-package duke.task;
+package datuk.task;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
